requirements/requirements.txt
Django==1.5.1
South==0.8.4
argparse==1.2.1
django-picklefield==0.3.1
psycopg2==2.5.2
six==1.5.2
wsgiref==0.1.2
